Move bolt configure process from checkout button click step to page load in order to make ApplePay work. And some code refactor.
